separate into utils file
